1. refine paddle parallel code in config.py 2. remove useless function in utils/paddle.py 3. move function split_in_rank from array_ops_compat.py into internal.py
