+++ deepxde/icbc/boundary_conditions.py
@@ -184,8 +184,8 @@ class PointSetBC:
     ):
         self.points = np.array(points, dtype=config.real(np))
         self.values = bkd.as_tensor(values, dtype=config.real(bkd.lib))
-        self.points = utils.array_ops_compat.split_in_rank(self.points)
-        self.values = utils.array_ops_compat.split_in_rank(self.values)
+        self.points = utils.split_in_rank(self.points)
+        self.values = utils.split_in_rank(self.values)
         self.component = component
         if isinstance(component, list) and backend_name != "pytorch":
             # TODO: Add support for multiple components in other backends
